requirements.txt
absl-py==1.3.0
addict==2.4.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
attrs==22.1.0
autopep8==2.0.2
Babel==2.11.0
backcall==0.2.0
basicsr==1.4.2
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
comm==0.1.2
contourpy==1.0.6
cupy-cuda112==10.6.0
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==1.2.0
facexlib==0.2.5
fastjsonschema==2.16.2
fastrlock==0.8.1
ffmpeg==1.4
filterpy==1.4.5
flatbuffers==23.5.9
fonttools==4.38.0
fqdn==1.5.1
future==0.18.2
gast==0.4.0
gfpgan==1.3.8
google-api-core==2.11.0
google-api-python-client==2.87.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.59.0
grpcio==1.51.1
h5py==3.8.0
httplib2==0.22.0
idna==3.4
imageio==2.23.0
importlib-metadata==5.1.0
ipykernel==6.19.2
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.3
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.10
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter_server==2.0.1
jupyter_server_terminals==0.4.2
jupyterlab==3.5.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.4
jupyterlab_server==2.16.5
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==16.0.0
llvmlite==0.39.1
lmdb==1.4.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.8
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.5
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==22.0
pafy==0.5.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.4
pytorch-msssim==0.2.1
pytorch-ssim==0.1
pytz==2022.6
PyWavelets==1.4.1
pywin32==305
pywinpty==2.0.9
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-image==0.19.3
scikit-video==1.1.11
scipy==1.9.3
Send2Trash==1.8.0
six==1.16.0
sk-video==1.1.10
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
tb-nightly==2.12.0a20230103
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
tifffile==2022.10.10
tinycss2==1.2.1
tomli==2.0.1
torch==1.13.1+cu116
torchvision==0.14.1+cu116
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
typing_extensions==4.4.0
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.13
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.4
wrapt==1.15.0
yapf==0.32.0
youtube-dl==2021.12.17
zipp==3.11.0
